updated commit

+++ src/dao/ProductDAOImpl.java
@@ -1,4 +1,3 @@
-
 package dao;
 
 import java.util.List;
@@ -8,18 +7,17 @@ import javax.persistence.Persistence;
 import javax.persistence.Query;
 import entities.Product;
 
+public class ProductDAOImpl implements ProductDAO {
 
+    EntityManagerFactory emf = Persistence.createEntityManagerFactory("PU");
+    //  
 
-public class ProductDAOImpl implements ProductDAO{
- EntityManagerFactory emf = Persistence.createEntityManagerFactory("PU");
-        
-    
     @Override
     public void create(String name, double price) {
         EntityManager em = emf.createEntityManager();
-        
+
         Product p = new Product(name, price);
-        
+
         em.getTransaction().begin();
         em.persist(p);
         em.getTransaction().commit();
@@ -29,10 +27,10 @@ public class ProductDAOImpl implements ProductDAO{
     @Override
     public List<Product> getByName(String name) {
         EntityManager em = emf.createEntityManager();
-       
+
         Query query = em.createNamedQuery("getByName", Product.class);
-        query.setParameter("name", "%"+name+"%");
-        List <Product>list = query.getResultList();
+        query.setParameter("name", "%" + name + "%");
+        List<Product> list = query.getResultList();
         em.close();
         return list;
     }
@@ -43,7 +41,7 @@ public class ProductDAOImpl implements ProductDAO{
         Query q = em.createNamedQuery("getByPriceRange", Product.class);
         q.setParameter("min", min);
         q.setParameter("max", max);
-        List <Product>list = q.getResultList();
+        List<Product> list = q.getResultList();
         em.close();
         return list;
     }
@@ -51,44 +49,59 @@ public class ProductDAOImpl implements ProductDAO{
     @Override
     public List<Product> getAll() {
         EntityManager em = emf.createEntityManager();
-        List <Product>list = em.createNamedQuery("getAll").getResultList();
+        List<Product> list = em.createNamedQuery("getAll").getResultList();
         em.close();
         return list;
     }
 
     @Override
-    public boolean updatePrice(int id, double newPrice) {
+    public void updatePrice(int id, double newPrice) {
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         Product p = em.find(Product.class, id);
-        p.setPrice(newPrice);
-        em.getTransaction().commit();
-        em.close();
-        return true; 
-        
+        if (p != null) {
+            p.setPrice(newPrice);
+            em.getTransaction().commit();
+            em.close();
+        }else{
+            System.out.println("Product was not found");
+        }
+
     }
 
     @Override
-    public boolean remove(int id) {
+    public void remove(int id) {
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         Product a = em.find(Product.class, id);
-        em.remove(a);
-        em.getTransaction().commit();
-        em.close();
-        return true; 
+
+        if (a != null) {
+            em.remove(a);
+            em.getTransaction().commit();
+            em.close();
+
+        } else {
+            System.out.println("The product doesn't exist");
+
+        }
+
     }
 
     @Override
-    public boolean updateName(int id, String name) {
+    public void updateName(int id, String name) {
         EntityManager em = emf.createEntityManager();
         em.getTransaction().begin();
         Product a = em.find(Product.class, id);
-        a.setName(name);
-        em.getTransaction().commit();
-        em.close();
-        return true;
+        if (a != null) {
+            a.setName(name);
+            em.getTransaction().commit();
+            em.close();
+
+        } else {
+            System.out.println("Not possible");
+
+        }
+
     }
 
-    
 }
